Backend

+++ frontend/components/product_index/user_product_item.jsx
@@ -1,44 +1,69 @@
-import React from 'react'
-import { Link } from 'react-router-dom'
+import React from 'react';
+import { Link } from 'react-router-dom';
+import { withRouter } from 'react-router-dom';
 
 class UserProductIdxItem extends React.Component{
     constructor(props) {
         super(props)
+        this.state = {
+            popup: false
+        }
 
-        this.showLess = this.showLess.bind(this)
-        this.imageTag = this.imageTag.bind(this)
+        this.showLess = this.showLess.bind(this);
+        this.imageTag = this.imageTag.bind(this);
+        this.keep = this.keep.bind(this);
     }
 
-    showLess(content) {
-        if (content.length > 16) {
+    showLess(product) {
+        let name = product.name;
+        if (name.length > 16) {
             return (
-                <h1 className="public-name">{content.slice(0, 16) + "..."}</h1>
+                <h1 className="public-name-link" onClick={() => this.props.history.push(`/products/${product.id}`)}>{name.slice(0, 16) + "..."}</h1>
             )
         } else {
             return (
-                <h1 className="public-name">{content}</h1>
+                <h1 className="public-name-link" onClick={() => this.props.history.push(`/products/${product.id}`)}>{name}</h1>
             )
         }
     }
 
     imageTag() {
         return (
-            <img className="idx-images"src={this.props.product.photoUrl} alt="" />
+            <img className="edit-idx-images"src={this.props.product.photoUrl} alt="" />
         )
     }
 
+    keep() {
+        let { product, deleteProduct } = this.props;
+        if (product.keep) {
+            setTimeout(this.clearPopUp, 5000)
+
+            return (
+                <div className="user-product-idx-item">
+                    {this.showLess(product)}
+                    {this.imageTag()}
+                    <button className="disable-link" onClick={() => this.showPopUp()}>Update Product Listing</button>
+                    <button className="disable-link" onClick={() => this.showPopUp()}>Delete Product</button>
+                    <div className="disable-div">Products used to seed the database have their delete and update features disabled! Feel free to create a product listing and try out these features! </div>
+                </div>
+            )
+        } else {
+            return (
+                <div className="user-product-idx-item">
+                    {this.showLess(product)}
+                    {this.imageTag()}
+                    <Link className="product-links" to={`/products/${product.id}/edit`}>Update Product Listing</Link>
+                    <button className="product-delete" onClick={() => deleteProduct(product.id)}>Delete Product</button>
+                </div>
+            )
+        }
+    }
+
     render() {
-        let {product, deleteProduct} = this.props
         return(
-            <div className="user-product-idx-item">
-                {this.showLess(product.name)}
-                {this.imageTag()}
-                <Link className="product-links" to={`/products/${product.id}/edit`}>Update Product Listing</Link>
-                <button className="product-delete" onClick={() => deleteProduct(product.id)}>Delete Product</button>
-            </div>
-            
+            this.keep()
         )
     }
 }
 
-export default UserProductIdxItem
+export default withRouter(UserProductIdxItem)
